+++ Enhetsregisteret/AT.Common.Enhetsregisteret.Publish/Ports/IEnhetsregisteret.cs
@@ -1,12 +1,16 @@
-using Arbeidstilsynet.Common.Enhetsregisteret.Model.Brreg;
 using Arbeidstilsynet.Common.Enhetsregisteret.Model.Request;
 using Arbeidstilsynet.Common.Enhetsregisteret.Model.Response;
+using Arbeidstilsynet.Common.Enhetsregisteret.Models;
 
 namespace Arbeidstilsynet.Common.Enhetsregisteret.Ports;
 
 /// <summary>
-/// Interface which can be dependency injected to use methods of Enhetsregisteret
+/// Interface describing common Enhetsregisteret operations.
 /// </summary>
+/// <remarks>
+/// This interface is not registered by <c>AddEnhetsregisteret(...)</c>. It is kept as a starting point
+/// for consumers who want to build their own adapter over the generated <c>EnhetsregisteretClient</c>.
+/// </remarks>
 public interface IEnhetsregisteret
 {
     /// <summary>
@@ -51,7 +55,7 @@ public interface IEnhetsregisteret
     /// <param name="query"></param>
     /// <param name="pagination"></param>
     /// <returns></returns>
-    Task<PaginationResult<Oppdatering>?> GetOppdateringerUnderenheter(
+    Task<PaginationResult<OppdateringerUnderenhet>?> GetOppdateringerUnderenheter(
         GetOppdateringerQuery query,
         Pagination pagination
     );
@@ -62,7 +66,7 @@ public interface IEnhetsregisteret
     /// <param name="query"></param>
     /// <param name="pagination"></param>
     /// <returns></returns>
-    Task<PaginationResult<Oppdatering>?> GetOppdateringerEnheter(
+    Task<PaginationResult<OppdateringerEnhet>?> GetOppdateringerEnheter(
         GetOppdateringerQuery query,
         Pagination pagination
     );